+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 
 ## [WIP]
 
-- Add support for websites: avbase.net, fc2ppvdb.com
+- Add support for websites: avbase.net, fc2ppvdb.com, duga.jp
 - Add aliases matching for tags, performers and studios
 - Add batchSize setting [#65]
 - Add query method setting (get/post) [#65]